(IAC-1233) Remove litmus as a dependency in Ruby 2.5 & 2.6
With the release of Puppet 7, we will soon run in to issues on our
CI pipelines for the Puppet supported modules, as bolt has quite
a narrow band of support for what version Puppet it supports.

There is no requirement for us to depend on Litmus -> Bolt for spec
test execution on Ruby 2.5 / Puppet 6 and acceptance tests can run
on any Ruby version.

This change removes Litmus as a dependency on Ruby 2.5 and 2.6 to
facilitate Puppet 5, 6 & 7 spec and acceptance test execution on the
supported modules CI pipelines.